back/requirements.txt
flask
flask-cors
langchain
langchain-google-genai
google-search-results
python-dotenv
gunicorn

